+++ monitor/CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- command line parameter `--db-min` and `--db-max` to set the dB range of the
+  spectrum and waterfall plot
+- ui config options `db_min` and `db_max` to set the dB range of the spectrum
+  and waterfall plot
 
 ## [0.3.0] - 2019-12-02
 ### Added